+++ CHANGELOG.md
@@ -52,6 +52,9 @@ Miscellaneous:
   automatically close the open project with the last stack-viewer, uncheck the
   "Project closes with last stack viewer" option in the Settings Widget.
 
+- Neuron similarity detail widget: make CSV export of object IDs and names
+  configurable.
+
 ### Bug fixes